api/requirements.txt
uvicorn 
fastapi 
python-dotenv 
openai 
numpy